Rename entities -> entity for clear imports
