+++ internal/core/entity/team.go
@@ -1,4 +1,4 @@
-package entities
+package entity
 
 import "github.com/google/uuid"
